+++ cmake/macros/IfuncTest.c
@@ -32,14 +32,26 @@ static int ifunc_method_impl3(int a, long b, char c)
 	return (int)(a * b + c) * 3;
 }
 
+static int ifunc_method_impl4(int a, long b, char c)
+{
+	return (int)(a * b + c) * 4;
+}
+
 static __typeof__(&ifunc_method_impl1) ifunc_method_resolve(void)
 {
+	// NOTE: Since libromdata is a shared library now, IFUNC resolvers
+	// cannot call PLT functions. Otherwise, it will crash.
+	// We'll use gcc's built-in CPU ID functions instead.
+	// Requires gcc-4.8 or later, or clang-6.0 or later.
 	time_t s = time(NULL);
-	switch (s % 3) {
+	__builtin_cpu_init();
+	if (__builtin_cpu_supports("sse2")) s ^= 3;
+	switch (s & 3) {
 		default:
 		case 0:	return &ifunc_method_impl1;
 		case 1:	return &ifunc_method_impl2;
 		case 2:	return &ifunc_method_impl3;
+		case 3:	return &ifunc_method_impl4;
 	}
 	return NULL;
 }
@@ -55,7 +67,7 @@ int main(void)
 	char c = 3;
 
 	int ret = ifunc_method(a, b, c);
-	if (ret == 38 || ret == 38*2 || ret == 38*3)
+	if (ret == 38 || ret == 38*2 || ret == 38*3 || ret == 38*4)
 		return EXIT_SUCCESS;
 	return EXIT_FAILURE;
 }